UDP Acknowledgement implemented
Since the other branch was changed quite a bit, this branch is revamped
for UDP stuff. Clients now send acknowledgement messages to whoever sent
them a message. This means with more than 2 clients many
Acknowledgements may be received. Each is tagged with the username and
implicitly the IP, though, so it is possible to display who did NOT
receive the message. Not implemented yet. Other features more important.